Made changes to classify and improved on psi calculations

+++ resources/isoformexpress/parseIsoform.py
@@ -4,68 +4,67 @@ import os,sys
 import csv
 import getopt
 import pybedtools
+import gzip
 from pandas import *
 
 ################################################################################
 # isoformExpression
-#"Brain-Nucleus_accumbens_basal_ganglia",
-#"Brain-Putamen_basal_ganglia",
-#"Brain-Spinal_cord_cervical_c-1",
-#"Brain-Substantia_nigra",
-#"Breast-Mammary_Tissue",
-#"Cells-EBV-transformed_lymphocytes",
-#"Cells-Leukemia_cell_line_CML",
-#"Cells-Transformed_fibroblasts",
-#"Colon-Transverse",
-#"Esophagus-Mucosa",
-#"Esophagus-Muscularis",
-#"Fallopian_Tube",
-#"Heart-Atrial_Appendage",
-#"Heart-Left_Ventricle",
-#"Kidney-Cortex",
-#"Liver","Lung",
-#"Muscle-Skeletal",
-#"Nerve-Tibial","Ovary","Pancreas",
-#"Pituitary","Stomach","Testis","Thyroid","Uterus",
-#"Vagina", "Whole_Blood"
-#"Adipose-Subcutaneous","Adipose-Visceral_Omentum",
-#"Adrenal_Gland","Artery-Aorta","Artery-Coronary",
-#"Artery-Tibial","Brain-Amygdala",
-#"Brain-Anterior_cingulate_cortex_BA24", 
-#"Brain-Caudate_basal_ganglia",
-#"Brain-Cerebellar_Hemisphere",
-
+AllTissues = [
+    "Brain-Nucleus_accumbens_basal_ganglia",
+    "Brain-Putamen_basal_ganglia",
+    "Brain-Spinal_cord_cervical_c-1",
+    "Brain-Substantia_nigra",
+    "Breast-Mammary_Tissue",
+    "Brain-Cerebellum",
+    "Brain-Cortex",
+    "Brain-Frontal_Cortex_BA9",
+    "Brain-Hippocampus",
+    "Brain-Hypothalamus",
+    "Colon-Transverse",
+    "Esophagus-Mucosa",
+    "Esophagus-Muscularis",
+    "Fallopian_Tube",
+    "Heart-Atrial_Appendage",
+    "Heart-Left_Ventricle",
+    "Kidney-Cortex",
+    "Liver","Lung",
+    "Muscle-Skeletal",
+    "Nerve-Tibial","Ovary","Pancreas",
+    "Pituitary","Stomach","Testis","Thyroid","Uterus",
+    "Vagina", "Whole_Blood",
+    "Adipose-Subcutaneous","Adipose-Visceral_Omentum",
+    "Adrenal_Gland","Artery-Aorta","Artery-Coronary",
+    "Artery-Tibial","Brain-Amygdala",
+    "Brain-Anterior_cingulate_cortex_BA24", 
+    "Brain-Caudate_basal_ganglia",
+    "Brain-Cerebellar_Hemisphere",
+]
+    #"Cells-EBV-transformed_lymphocytes",
+    #"Cells-Leukemia_cell_line_CML",
+    #"Cells-Transformed_fibroblasts",
 #
 ################################################################################
-def isoformExpression( isoformfile, generef, force=False ) :
-    isodict = read_csv( "rawdata/CG_named_loci.tsv", sep="\t" )
-    isodict.index= isodict["IsoformID"]
+def isoformExpression( isoformfile, generef, tissuelist, force=False ) :
+    #isodict = read_csv( "rawdata/CG_named_loci.tsv", sep="\t" )
+    #isodict.index= isodict["IsoformID"]
 
     if os.path.exists(isoformfile) and not force : 
         genedata = read_csv(isoformfile, sep="\t",index_col="GeneSymbol")
         return genedata
 
-    tissuelist = [
-                  "Brain-Cerebellum",
-                  "Brain-Cortex"
-                  #"Brain-Frontal_Cortex_BA9",
-                  #"Brain-Hippocampus"
-                  #"Brain-Hypothalamus"
-                ]
-
     #tissue = "Brain-Cerebellar_Hemisphere"
     #targetfile = "rawdata/Brain-Cerebellar_Hemisphere_fpkm.tsv"
     isodata = {}
     for tissue in tissuelist :
-        targetfile = "rawdata/%s_fpkm.tsv" % tissue
+        targetfile = "rawdata/%s_fpkm.tsv.gz" % tissue
         print targetfile, os.path.exists(targetfile)
         limit = -1
         cdata = {}
-        FH = open(targetfile)
+        FH = gzip.open(targetfile)
         for row in csv.reader(FH,delimiter="\t"):
             if row[0] == "IsoformID": continue
             isoid = row[0]
-            meanscore = sum([float(x) for x in row[1:]])/len(row)
+            meanscore = sum([float(x.replace(",",".")) for x in row[1:]])/len(row)
             #print isoid, meanscore
             cdata[isoid] = meanscore
             limit -= 1
@@ -101,10 +100,11 @@ def intersectRanges( grange, iranges ):
 # END intersectRanges
 
 def parseIsoformRegions( targetfile, generef, force=False, limit=-1):
-    print "Running parseIsoformRegions"
+    print "Running parseIsoformRegions", targetfile
     regionfile = "CG.gtf"
 
     if os.path.exists(targetfile) and not force : 
+        print "Warning: isoformfile already exists. skipping..."
         gtfdata = read_csv(targetfile, sep="\t")
         return gtfdata
 
@@ -123,26 +123,19 @@ def parseIsoformRegions( targetfile, generef, force=False, limit=-1):
     return gtfdata
 # END parseIsoformRegions
 
-################################################################################
-# calculateRelativeExpression
-################################################################################
-def calculateRelativeExpression( exprfile ):
-    print "Running calculateRelativeExpression"
-    generef = read_csv("CG_named_loci.tsv",sep="\t")
-    generef.columns = [x.strip() for x in generef.columns]
-
-    targetfile = "results/isoexpr.tsv"
-    exprdata = isoformExpression(targetfile, generef, force=True)
-
-    gtfsummaryfile = "results/gtfsummary.tsv"
-    gtfdata = parseIsoformRegions(gtfsummaryfile, generef, force=True)
+def makeExpressionFile( exprdata, gtfdata, targetdir, tissue, force=False ):
+    print "Running makeExpressionFile:",tissue
+    exprfile = "%s/%s_expr.txt" % (targetdir, tissue)
+    if os.path.exists( exprfile ) and not force :
+        print "Warning: Tissue found already",tissue,"skipping..."
+        return exprfile
 
     OUT = open( exprfile , "w" )
-    OUT.write("chrom\tstart\tend\tGene\tisocov\tisocount\trelexpr\ttotalexpr\tIsoforms\n")
+    OUT.write("chrom\tstart\tend\tGene\tisocov\tisocount\tpsi\tsubexpr\ttotalexpr\tIsoforms\n")
     for index, irange in gtfdata.groupby(["GeneSymbol","chrom"]) :
         gene,chrom = index
         geneexpr = DataFrame({"IsoformID":exprdata.ix[gene]["IsoformID"].split(';'),
-                             'Expr':exprdata.ix[gene]["Brain-Cerebellum"].split(';')})
+                             'Expr':exprdata.ix[gene][tissue].split(';')})
         isoforms = irange.IsoformID.unique()
         geneexpr = geneexpr[geneexpr.IsoformID.isin(isoforms)]
         totalexpr = sum(geneexpr.Expr.map(float))
@@ -152,18 +145,75 @@ def calculateRelativeExpression( exprfile ):
         for i in range(len(generanges)) :
             if covdf.ix[i].RangeCount == 0 : continue
             isoset = covdf.ix[i].Isoset
-            subexpr = (geneexpr[geneexpr.IsoformID.isin(isoset)].Expr.map(float))
-            relexpr = sum(subexpr)/totalexpr if totalexpr > 0 else 1
-            s = ("%s\t%d\t%d\t%s\t%d\t%d\t%.3f\t%.3f\t%s\n" %
+            subexpr = sum(geneexpr[geneexpr.IsoformID.isin(isoset)].Expr.map(float))
+            psi = "%.1f" % (subexpr/totalexpr*100) if totalexpr > 0 else "-"
+            s = ("%s\t%d\t%d\t%s\t%d\t%d\t%s\t%.3f\t%.3f\t%s\n" %
                              (chrom, int(generanges[i]),
                               int(generanges[i+1]),gene, 
                               covdf.ix[i].RangeCount, 
                               len(isoforms), 
-                              relexpr, totalexpr,
+                              psi, subexpr, totalexpr,
                               ",".join(covdf.ix[i].Isoset)) )
             OUT.write( s )
 
     OUT.close()
+    return exprfile
+# END makeExpressionFile
+
+################################################################################
+# calculateRelativeExpression
+################################################################################
+def calculateRelativeExpression( targetdir="results" ):
+    print "Running calculateRelativeExpression"
+    generef = read_csv("CG_named_loci.tsv",sep="\t")
+    generef.columns = [x.strip() for x in generef.columns]
+
+    tissuelist = [
+                  "Brain-Cerebellum",
+                  "Brain-Cortex"
+                  #"Brain-Frontal_Cortex_BA9",
+                  #"Brain-Hippocampus"
+                  #"Brain-Hypothalamus"
+                ]
+    #tissuelist = AllTissues 
+
+    targetfile = "results/isoexpr.tsv"
+    exprdata = isoformExpression(targetfile, generef, tissuelist, force=False)
+
+    gtfsummaryfile = os.path.join(targetdir,"gtfsummary.tsv")
+    gtfdata = parseIsoformRegions(gtfsummaryfile, generef, force=False)
+
+    allfiles = {}
+    tissueexprdir = os.path.join(targetdir,"tissueexpr")
+    for tissue in tissuelist : 
+        cfile = makeExpressionFile( exprdata, gtfdata, tissueexprdir, tissue )
+        allfiles[tissue] = cfile
+                                   
+    print allfiles
+    alldata = None
+    psifile = os.path.join(targetdir,"psi_results.txt")
+    for tissue in allfiles : 
+        print tissue, allfiles[tissue]
+        tdata = read_csv( allfiles[tissue], sep="\t" )
+        print tdata.head()
+        if alldata is None :
+            alldata = tdata[["chrom","start","end","Gene","isocov","isocount"]]
+
+        alldata[tissue] = tdata["psi"]
+
+    print alldata[allfiles.keys()].head()
+    #alldata["Global-PSI"] = "-"
+    psi = []
+    for idx, row in alldata[allfiles.keys()].iterrows() :
+        subrow = [float(x) for x in row if x != "-"]
+        if len(subrow) == 0 : psi.append("-")
+        else : psi.append(sum(subrow)/len(subrow))
+
+    alldata["Global-PSI"] = psi
+    #alldata["Global-PSI"] = alldata[allfiles.keys()].mean(0,numeric_only=True)
+    print alldata.head()
+    alldata.to_csv(psifile, sep="\t", index=False )
+
 # END calculateRelativeExpression
 
 ################################################################################
@@ -181,7 +231,7 @@ if __name__ == "__main__" :
     testexprfile = "results/relativeexpr1.bed"
     exprfile = "results/relativeexpr.bed"
     if optlist.has_key("-c") : 
-        calculateRelativeExpression(exprfile)
+        calculateRelativeExpression()
     else: 
         query = optlist.get("-s",None)
         assert query.find(":") > 0, "Error: incorrect formatted query"+ query
